captcha_solver/requirements.txt
tensorflow
keras
numpy
pandas
opencv-python
pillow
matplotlib
scikit-learn
paddleocr
paddlepaddle